# Simplify and optimize PlayIterator.is_failed