--- a/changelogs/fragments/playiterator-is_failed-optimize.yml
+++ b/changelogs/fragments/playiterator-is_failed-optimize.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - "PlayIterator - optimize the 'is host failed' check."