##### SUMMARY
There should not be a need for the complicated logic. We can say the host is failed only after it has completed and there was a failure. The completed part is important as a host can have a failure set on it but there could still be tasks to run in the rescue or always sections.

Perf test:
Iterating through N nested blocks, each block has one task and a nested block, failing the most nested block.

Before:
0.0002570152282714844 seconds, 10 blocks
0.004849910736083984 seconds, 50 blocks
0.9955852031707764 seconds, 350 blocks
21.4445219039917 seconds, 992 blocks

After:
0.00022101402282714844 seconds, 10 blocks
0.0024309158325195312 seconds, 50 blocks
0.1517040729522705 seconds, 350 blocks
1.9403960704803467 seconds, 992 blocks

ci_complete
<!--- Describe the change below, including rationale and design decisions -->

<!--- Add "Fixes #1234" or steps to reproduce the problem if there is no corresponding issue -->

##### ISSUE TYPE

<!--- Pick one below and delete the rest -->
- Feature Pull Request